+++ changelog.md
@@ -7,6 +7,8 @@
 
 #### v10.02
 
+- Renamed media server to kamehouse-r2d2
+- Added vlc stats view to ui in debug mode
 - Added patch script to patch other kamehouse servers with current working copy
 - Fixed vlc stats for mp3
 - Fixed bugs